python-ai-service/requirements.txt
# Python AI Service Requirements
# Core Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# AI & Machine Learning
openai==1.3.5
google-generativeai==0.3.1
langchain==0.0.350
langchain-openai==0.0.2
anthropic==0.7.7

# Data Processing
pandas==2.1.3
numpy==1.24.3

# Database & Storage
pymongo==4.6.0
motor==3.3.2
redis==5.0.1

# HTTP & API
httpx==0.25.2
requests==2.31.0

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic-settings==2.1.0

# Date & Time
python-dateutil==2.8.2

# Logging & Monitoring
loguru==0.7.2

# CORS
fastapi-cors==0.0.6
